requirements.txt
textual>=0.47.0
google-genai
Pillow>=9.0.0
python-dotenv>=1.0.0
